+++ lib/fading_scroll.dart
@@ -1,3 +1,1 @@
-library fading_scroll;
-
 export 'src/presentation/fading_scroll.dart';